+++ FrameworkQ.ConsularServices/Users/RoleUserMap.cs
@@ -5,8 +5,8 @@ namespace FrameworkQ.ConsularServices.Users;
 public class RoleUserMap
 {
     [Column("user_id")]
-    public int UserId { get; set; }
+    public long UserId { get; set; }
 
     [Column("role_id")]
-    public int RoleId { get; set; }
+    public long RoleId { get; set; }
 }
